split test_phonons.py into test_helpers.py and test_plotly.py
