+++ tests/files/phonons/mp-2667-Cs1Au1-chgnet-v0.3.0.json.xz
