+++ tests/files/phonons/mp-2667-Cs1Au1-m3gnet.json.xz
