+++ tests/files/phonons/mp-2691-Cd4Se4-chgnet-v0.3.0.json.xz
